Fix warnings, turn on -ansi -pedantic
The Makefule now usies -ansi -pedantic when compiling all the code.
This forces strict compliance with the C99 specification. Some cleanups
have been made to silence all the warnings this exposed. A few warnings
remain, but I'm going to hold off on fixing them until I can actually
test the code on hardware to make sure nothing breaks.

Next step will also be to turn on -Wconversion. Testing with this
exposed even more warnings, which will be fixed in the next update.

+++ os/rt/ports/ARMCMx/chcore_v6m.c
@@ -131,12 +131,12 @@ void _port_irq_epilogue(regarm_t lr) {
        required or not.*/
     if (chSchIsPreemptionRequired()) {
       /* Preemption is required we need to enforce a context switch.*/
-      ctxp->pc = (regarm_t)_port_switch_from_isr;
+      ctxp->pc = (regarm_t)(uint32_t)_port_switch_from_isr;
     }
     else {
       /* Preemption not required, we just need to exit the exception
          atomically.*/
-      ctxp->pc = (regarm_t)_port_exit_from_isr;
+      ctxp->pc = (regarm_t)(uint32_t)_port_exit_from_isr;
     }
 
     /* Note, returning without unlocking is intentional, this is done in
